Detalles con la BD y agregación de nuevas vistas
Se arreglaron detalles en el backend (API), se agrego la vista del carrito de compras, se modifico la vista de admin y ahora esta todo funcional en esta, ahora se pueden crear cuentas y pueden acceder a ellas mismas, se implementaron funciones con Tokens para los inicios de sesion para los administradores y los usuarios

+++ src/views/VistaAdmin.vue
@@ -1,6 +1,6 @@
 <template>
     <v-app>
-        <v-row class="mx-2 mt-3">
+        <v-row class="mx-2 mt-3" v-if="ingresoAdmin === true">
             <!-- Dialogo de agregar -->
             <v-dialog v-model="drawerAgregar" max-width="500px">
                 <v-card>
@@ -27,8 +27,8 @@
                                         :rules="[() => !!precio || 'Este campo no puede estar vacio']" label="Precio"
                                         required>
                                     </v-text-field>
-                                    <v-text-field v-model="talla"
-                                        :rules="[() => !!talla || 'Este campo no puede estar vacio']" label="Tallas"
+                                    <v-text-field v-model="stock"
+                                        :rules="[() => !!stock || 'Este campo no puede estar vacio']" label="stocks"
                                         required>
                                     </v-text-field>
                                     <v-text-field v-model="color"
@@ -75,9 +75,16 @@
                                         :rules="[() => !!precio || 'Este campo no puede estar vacio']" label="Precio"
                                         required>
                                     </v-text-field>
-                                    <v-text-field v-model="talla"
-                                        :rules="[() => !!talla || 'Este campo no puede estar vacio']" label="Tallas"
-                                        required>
+                                    <p>
+                                        Tallas:
+                                    </p>
+                                    <li v-for="(product) in stock" :key="product.id">
+                                        <input :id="product.id" :value="product.talla" name="product.talla"
+                                            type="checkbox" v-model="checked">
+                                        <label :for="product.id"> {{product.talla}}</label>
+                                    </li>
+                                    <v-text-field :rules="[() => !!cantidad || 'Este campo no puede estar vacio']"
+                                        label="Cantidad" required>
                                     </v-text-field>
                                     <v-text-field v-model="color"
                                         :rules="[() => !!color || 'Este campo no puede estar vacio']" label="Color"
@@ -99,59 +106,76 @@
             </v-dialog>
             <!-- Mostrar los datos de las zapatillas del inventario -->
             <v-row class="mx-12 mt-4" justify="center">
-                <v-container fluid style="margin: 0px; padding: 0px; width: 35%">
-                    <v-row class="mb-12" v-for="zapatilla in zapatillas" :key="zapatilla._id">
-                        <v-img height="180px" width="160px" :src="zapatilla.imagen">
-                            <template v-slot:placeholder>
-                                <v-row class="fill-height ma-0" align="center" justify="center">
-                                    <v-progress-circular indeterminate color="teal lighten-2"></v-progress-circular>
-                                </v-row>
-                            </template>
-                        </v-img>
-                        <v-col>
-                            <h1 style="font-size:170%">{{zapatilla.nombre}}</h1>
-                            <p style="font-size:90%">
-                                Marca: {{zapatilla.marca}}<br>
-                                Tipo: {{zapatilla.tipo}} <br>
-                                Precio: {{zapatilla.precio}}<br>
-                                Tallas: {{zapatilla.talla}}<br>
-                                Color: {{zapatilla.color}}<br>
-                                Id: {{zapatilla._id}}<br>
-                            </p>
+                <v-container fluid style="margin: 0px; padding: 0px; width: 60%">
+                    <v-row>
+                        <v-col v-for="zapatilla in zapatillas" :key="zapatilla._id" cols="3">
+                            <v-card height="550" width="700" outlined center>
+                                <v-img height="160" width="300" :src="zapatilla.imagen">
+                                    <template v-slot:placeholder>
+                                        <v-row class="fill-height ma-0" align="center" justify="center">
+                                            <v-progress-circular indeterminate color="teal lighten-2">
+                                            </v-progress-circular>
+                                        </v-row>
+                                    </template>
+                                </v-img>
+                                <v-col>
+                                    <v-btn class="mx-3" color="teal lighten-2" fab
+                                        @click="eliminarProducto(zapatilla._id)" top left card>
+                                        <Icon icon="fluent:delete-16-regular" width="30" height="30" />
+                                    </v-btn>
+                                    <v-btn color="teal lighten-2" fab
+                                        @click="editarProducto(true, zapatilla.nombre, zapatilla.marca, zapatilla.stock, zapatilla.color, zapatilla.precio, zapatilla.tipo, zapatilla._id, zapatilla.imagen)"
+                                        top right card>
+                                        <Icon icon="clarity:edit-solid" width="30" height="30" />
+                                    </v-btn>
+                                    <h1 style="font-size:140%">{{zapatilla.nombre}}</h1>
+                                    <p style="font-size:90%">
+                                        Marca: {{zapatilla.marca}}<br>
+                                        Tipo: {{zapatilla.tipo}} <br>
+                                        Precio: {{zapatilla.precio}}<br>
+                                        Tallas:<br>
+                                        <span v-for="{talla, cantidad} in zapatilla.stock" :key="talla.id">
+                                            talla: {{talla}} quedan: {{cantidad}} <br>
+                                        </span>
+                                        Color: {{zapatilla.color}}<br>
+                                        Id: {{zapatilla._id}}<br>
+                                    </p>
+                                </v-col>
+                            </v-card>
                         </v-col>
-                        <v-btn class="mx-3" color="teal lighten-2" fab @click="eliminarProducto(zapatilla._id)">
-                            <Icon icon="fluent:delete-16-regular" width="30" height="30" />
-                        </v-btn>
-                        <v-btn color="teal lighten-2" fab
-                            @click="editarProducto(true, zapatilla.nombre, zapatilla.marca, zapatilla.talla, zapatilla.color, zapatilla.precio, zapatilla.tipo, zapatilla._id, zapatilla.imagen)">
-                            <Icon icon="clarity:edit-solid" width="30" height="30" />
-                        </v-btn>
                     </v-row>
                 </v-container>
             </v-row>
             <!-- Boton agregar -->
             <div style: width="30px">
                 <v-row no-gutters justify="end">
-                    <v-btn @click="agregarProducto(true)" fab>
+                    <v-btn @click="agregarProducto(true)" absolute top right>
                         <Icon icon="carbon:add-filled" color="#4db6ac" width="66" height="66" />
                     </v-btn>
                 </v-row>
             </div>
         </v-row>
+        <v-container align="center" v-else>
+            <v-spacer></v-spacer>
+            <h1> No se encontró la página</h1>
+            <v-spacer></v-spacer>
+        </v-container>
     </v-app>
 
 </template>
 <script>
+import Swal from 'sweetalert2'
 import { Icon } from "@iconify/vue2";
 export default {
     data: () => ({
+        ingresoAdmin: null,
         name: null,
         marca: null,
         tipo: null,
         precio: null,
-        talla: null,
         color: null,
         imagen: null,
+        stock: [],
         _id: null,
         drawerAgregar: false,
         drawerEditar: false,
@@ -161,7 +185,7 @@ export default {
             marca: null,
             tipo: null,
             precio: null,
-            talla: null,
+            stock: null,
             color: null,
             imagen: null,
         },
@@ -170,23 +194,31 @@ export default {
             marca: null,
             tipo: null,
             precio: null,
-            talla: null,
+            stock: null,
             color: null,
             imagen: null,
             _id: null
         },
-
-        //rules: [
-        //value => !value || value.size < 2000000 || '¡La imagen no puede pesar mas de 2MB!',
-        //],
+        admins: [],
     }),
     created() {
         this.listarZapatillas();
+        this.listarCuentas();
+        this.revisarAdmin();
     },
     components: {
         Icon,
     },
     methods: {
+        listarCuentas() {
+            this.axios.get("EZ-Admin")
+                .then((response) => {
+                    this.admins = response.data;
+                })
+                .catch((e) => {
+                    console.log('error' + e);
+                })
+        },
         listarZapatillas() {
             this.axios.get('EZ-Producto')
                 .then((response) => {
@@ -201,7 +233,7 @@ export default {
                 this.drawerAgregar = true;
                 this.name = ''
                 this.marca = ''
-                this.talla = ''
+                this.stock = ''
                 this.tipo = ''
                 this.color = ''
                 this.precio = 0
@@ -211,13 +243,13 @@ export default {
                     marca: null,
                     tipo: null,
                     precio: null,
-                    talla: null,
+                    stock: null,
                     color: null,
                     imagen: null,
                 }
                 this.productoAgregar.nombre = this.name
                 this.productoAgregar.marca = this.marca
-                this.productoAgregar.talla = this.talla
+                this.productoAgregar.stock = this.stock
                 this.productoAgregar.tipo = this.tipo
                 this.productoAgregar.color = this.color
                 this.productoAgregar.precio = this.precio
@@ -233,21 +265,39 @@ export default {
             }
         },
         eliminarProducto(id) {
-            this.axios.delete(`Producto-el/${id}`)
-                .then(res => {
-                    let index = this.zapatillas.findIndex(item => item._id === res.data._id)
-                    this.zapatillas.splice(index, 1);
-                })
-                .catch(e => {
-                    console.log("error " + e.response);
-                })
+            Swal.fire({
+                title: '¿Estas seguro?',
+                text: "No podras revertir esto!",
+                icon: 'warning',
+                showCancelButton: true,
+                confirmButtonColor: '#3085d6',
+                cancelButtonColor: '#d33',
+                confirmButtonText: 'Si, eliminalo!'
+            }).then((result) => {
+                if (result.isConfirmed) {
+                    this.axios.delete(`Producto-el/${id}`)
+                        .then(res => {
+                            let index = this.zapatillas.findIndex(item => item._id === res.data._id)
+                            this.zapatillas.splice(index, 1);
+                        })
+                        .catch(e => {
+                            console.log("error " + e.response);
+                        })
+                    Swal.fire(
+                        'Eliminado!',
+                        'El archivo a sido eliminado.',
+                        'success'
+                    )
+                }
+            })
+
         },
-        editarProducto(consulta, name, marca, talla, color, precio, tipo, _id, imagen) {
+        editarProducto(consulta, name, marca, stock, color, precio, tipo, _id, imagen) {
             if (consulta == true) {
                 this.drawerEditar = true
                 this.name = name
                 this.marca = marca
-                this.talla = talla
+                this.stock = stock
                 this.tipo = tipo
                 this.color = color
                 this.precio = precio
@@ -255,7 +305,7 @@ export default {
                 this.imagen = imagen
             } else {
                 this.drawerEditar = false
-                if (this.name == '' || this.tipo == '' || this.marca == '' || this.talla == '' || this.color == '' || this.precio == '' || this.imagen == '') {
+                if (this.name == '' || this.tipo == '' || this.marca == '' || this.stock == [] || this.color == '' || this.precio == '' || this.imagen == '') {
                     console.log("Datos vacios")
                 } else {
                     this.productoEditar = {
@@ -263,14 +313,14 @@ export default {
                         marca: null,
                         tipo: null,
                         precio: null,
-                        talla: null,
+                        stock: null,
                         color: null,
                         imagen: null,
                         _id: null,
                     }
                     this.productoEditar.nombre = this.name
                     this.productoEditar.marca = this.marca
-                    this.productoEditar.talla = this.talla
+                    this.productoEditar.stock = this.stock
                     this.productoEditar.tipo = this.tipo
                     this.productoEditar.color = this.color
                     this.productoEditar.precio = this.precio
@@ -281,7 +331,7 @@ export default {
                             const index = this.zapatillas.findIndex(zapatillaB => zapatillaB._id === this.productoEditar._id);
                             this.zapatillas[index].nombre = this.productoEditar.nombre
                             this.zapatillas[index].marca = this.productoEditar.marca
-                            this.zapatillas[index].talla =this.productoEditar.talla
+                            this.zapatillas[index].stock = this.productoEditar.stock
                             this.zapatillas[index].tipo = this.productoEditar.tipo
                             this.zapatillas[index].color = this.productoEditar.color
                             this.zapatillas[index].precio = this.productoEditar.precio
@@ -292,6 +342,27 @@ export default {
                         })
                 }
             }
+        },
+        revisarAdmin() {
+            var estado = false
+            this.axios.get("EZ-Admin")
+                .then((response) => {
+                    this.admins = response.data;
+                    for (var i = 0; i < this.admins.length; i++) {
+                        console.log(this.admins[i].nombreUsuario + "???")
+                        if (localStorage.getItem(this.admins[i].nombreUsuario)) {
+                            estado = true
+                        }
+                    }
+                    if (estado === true) {
+                        this.ingresoAdmin = true
+                    } else {
+                        this.ingresoAdmin = false
+                    }
+                })
+                .catch((e) => {
+                    console.log('error' + e);
+                })
         }
     }
 }
